move components into specific word folder
