+++ components/WordComponents/WordComponent.tsx
